Final changes. We had alpha as an int instead of double. Some tests may differ on time but not something terrible.

+++ header_files/vamana.h
@@ -16,4 +16,4 @@ using namespace std;
 int medoid(vector<vector<float>>&vec);
 
 //implements vamana algorithm
-map<int,list<int>> vamana_index_algorithm(vector<vector<float>>&vec, int &R,int& medoid_node,int &L_sizelist, int& a,vector<vector<double>>& vecmatrix );
+map<int,list<int>> vamana_index_algorithm(vector<vector<float>>&vec, int &R,int& medoid_node,int &L_sizelist, double& a,vector<vector<double>>& vecmatrix );
